--- client/src/components/Login/index.js
@@ -1,1 +1,0 @@
-export { Login } from './Login'
