--- client/src/components/LoginScreen/index.js
@@ -1,1 +1,0 @@
-export { LoginScreen } from './LoginScreen'
